finished edit cabin

+++ src/services/apiCabins.js
@@ -1,5 +1,4 @@
-import supabase from './supabase';
-import { supabaseUrl } from './supabase';
+import supabase, { supabaseUrl } from './supabase';
 
 export async function getCabins() {
 	const { data, error } = await supabase.from('cabins').select('*');
@@ -12,19 +11,31 @@ export async function getCabins() {
 	return data;
 }
 
-export async function createCabin(newCabin) {
-	const imageName = `${Math.random()}-${newCabin.image.name}.jpg`.replaceAll(
-		'/',
-		''
-	);
+export async function createEditCabin(newCabin, id) {
+	console.log(newCabin, id);
 
-	const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+	const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-	console.log('newCabin', newCabin);
-	const { data, error } = await supabase
-		.from('cabins')
-		.insert([{ ...newCabin, image: imagePath }])
-		.select();
+	const imageName = hasImagePath
+		? newCabin.image
+		: `${Math.random()}-${newCabin.image.name}.jpg`.replaceAll('/', '');
+
+	const imagePath = hasImagePath
+		? imageName
+		: `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+	let query = supabase.from('cabins');
+
+	//CREATE
+	if (!id) {
+		query = query.insert([{ ...newCabin, image: imagePath }]);
+	}
+	//EDIT
+	else if (id) {
+		query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+	}
+
+	const { data, error } = await query.select().single();
 
 	if (error) {
 		console.error(error);
@@ -35,7 +46,7 @@ export async function createCabin(newCabin) {
 		.from('cabin-images')
 		.upload(imageName, newCabin.image);
 
-	if (storageError) {
+	if (!hasImagePath && storageError) {
 		await supabase.from('cabins').delete().eq('id', data.id);
 		console.error(storageError);
 		throw new Error(
